// src/generated/graphql.ts
export type Maybe<T> = T | null
// ...
export type AnalyticsPartnerStats = {
  __typename?: "AnalyticsPartnerStats"
  artworkPublished?: Maybe<AnalyticsPartnerStatsArtworksPublished>
  /** @deprecated Use artworkPublished for refactored time series bucket code */
  artworksPublished?: Maybe<AnalyticsArtworksPublishedStats>
  audience?: Maybe<AnalyticsPartnerAudienceStats>
  groupedStats?: Maybe<AnalyticsGroupedStatsConnection>
  inquiry?: Maybe<AnalyticsPartnerInquiryStats>
  pageview?: Maybe<AnalyticsPartnerStatsPageviews>
  /** @deprecated Use pageview for refactored time series bucket code */
  pageviews?: Maybe<AnalyticsPageviewStats>
  partnerId: Scalars["String"]["output"]
  rankedStats?: Maybe<AnalyticsRankedStatsConnection>
  sales?: Maybe<AnalyticsPartnerSalesStats>
  /** @deprecated Use rankedStats(objectType: ) instead */
  topArtworks?: Maybe<AnalyticsRankedStatsConnection>
  /** @deprecated Use audience() { uniqueVisitors } instead */
  uniqueVisitors?: Maybe<Scalars["Int"]["output"]>
}
// ...
export type AnalyticsRankedStats = {
  __typename?: "AnalyticsRankedStats"
  entity?: Maybe<AnalyticsRankedEntityUnion>
  period: AnalyticsQueryPeriodEnum
  rankedEntity: AnalyticsRankedStatsUnion
  value: Scalars["Int"]["output"]
}

export type AnalyticsRankedStatsConnection = {
  __typename?: "AnalyticsRankedStatsConnection"
  edges?: Maybe<Array<Maybe<AnalyticsRankedStatsEdge>>>
  nodes?: Maybe<Array<Maybe<AnalyticsRankedStats>>>
  pageInfo: AnalyticsPageInfo
}

export type AnalyticsRankedStatsEdge = {
  __typename?: "AnalyticsRankedStatsEdge"
  cursor: Scalars["String"]["output"]
  node?: Maybe<AnalyticsRankedStats>
}
// ...
export type Artist = EntityWithFilterArtworksConnectionInterface &
  Node &
  Searchable & {
    __typename?: "Artist"
    alternateNames?: Maybe<Array<Maybe<Scalars["String"]["output"]>>>
    articlesConnection?: Maybe<ArticleConnection>
    artistSeriesConnection?: Maybe<ArtistSeriesConnection>
    artworksConnection?: Maybe<ArtworkConnection>
    auctionResultsConnection?: Maybe<AuctionResultConnection>
    awards?: Maybe<Scalars["String"]["output"]>
    basedOn?: Maybe<Artist>
    biennials?: Maybe<Scalars["String"]["output"]>
    bio?: Maybe<Scalars["String"]["output"]>
    biography?: Maybe<Article>
    biographyBlurb?: Maybe<ArtistBlurb>
    birthday?: Maybe<Scalars["String"]["output"]>
    blurb?: Maybe<Scalars["String"]["output"]>
    cached?: Maybe<Scalars["Int"]["output"]>
    careerHighlights: Array<CareerHighlight>
    carousel?: Maybe<ArtistCarousel>
    collections?: Maybe<Array<Maybe<Scalars["String"]["output"]>>>
    contemporary?: Maybe<Array<Maybe<Artist>>>
    counts?: Maybe<ArtistCounts>
    coverArtwork?: Maybe<Artwork>
    createdAt?: Maybe<Scalars["String"]["output"]>
    criticallyAcclaimed: Scalars["Boolean"]["output"]
    currentEvent?: Maybe<CurrentEvent>
    deathday?: Maybe<Scalars["String"]["output"]>
    disablePriceContext?: Maybe<Scalars["Boolean"]["output"]>
    displayLabel?: Maybe<Scalars["String"]["output"]>
    displayName?: Maybe<Scalars["String"]["output"]>
    duplicates?: Maybe<Array<Maybe<Artist>>>
    exhibitionHighlights?: Maybe<Array<Maybe<Show>>>
    filterArtworksConnection?: Maybe<FilterArtworksConnection>
    first?: Maybe<Scalars["String"]["output"]>
    formattedArtworksCount?: Maybe<Scalars["String"]["output"]>
    formattedNationalityAndBirthday?: Maybe<Scalars["String"]["output"]>
    foundations?: Maybe<Scalars["String"]["output"]>
    gender?: Maybe<Scalars["String"]["output"]>
    genes?: Maybe<Array<Maybe<Gene>>>
    groupIndicator?: Maybe<ArtistGroupIndicator>
    hasMetadata?: Maybe<Scalars["Boolean"]["output"]>
    highlights?: Maybe<ArtistHighlights>
    hometown?: Maybe<Scalars["String"]["output"]>
    href?: Maybe<Scalars["String"]["output"]>
    id: Scalars["ID"]["output"]
    image?: Maybe<Image>
    imageUrl?: Maybe<Scalars["String"]["output"]>
    initials?: Maybe<Scalars["String"]["output"]>
    insights: Array<ArtistInsight>
    internalID: Scalars["ID"]["output"]
    isConsignable?: Maybe<Scalars["Boolean"]["output"]>
    isDisplayAuctionLink?: Maybe<Scalars["Boolean"]["output"]>
    isFollowed?: Maybe<Scalars["Boolean"]["output"]>
    isInSeoExperiment?: Maybe<Scalars["Boolean"]["output"]>
    isPersonalArtist?: Maybe<Scalars["Boolean"]["output"]>
    isPublic: Scalars["Boolean"]["output"]
    isShareable?: Maybe<Scalars["Boolean"]["output"]>
    last?: Maybe<Scalars["String"]["output"]>
    location?: Maybe<Scalars["String"]["output"]>
    marketingCollections: Array<MarketingCollection>
    meta: ArtistMeta
    middle?: Maybe<Scalars["String"]["output"]>
    name?: Maybe<Scalars["String"]["output"]>
    nationality?: Maybe<Scalars["String"]["output"]>
    partnerArtists?: Maybe<Array<Maybe<PartnerArtist>>>
    /** @deprecated This field is deprecated. No longer in use */
    partnerBiographyBlurb?: Maybe<PartnerBiographyBlurb>
    partnersConnection?: Maybe<PartnerArtistConnection>
    recentShow?: Maybe<Scalars["String"]["output"]>
    related?: Maybe<ArtistRelatedData>
    reviewSources?: Maybe<Scalars["String"]["output"]>
    sales?: Maybe<Array<Maybe<Sale>>>
    showsConnection?: Maybe<ShowConnection>
    slug: Scalars["ID"]["output"]
    sortableID?: Maybe<Scalars["String"]["output"]>
    statuses?: Maybe<ArtistStatuses>
    targetSupply: ArtistTargetSupply
    vanguardYear?: Maybe<Scalars["String"]["output"]>
    verifiedRepresentatives: Array<VerifiedRepresentative>
    years?: Maybe<Scalars["String"]["output"]>
  }
// ...
export type CroppedImageUrl = {
  __typename?: "CroppedImageUrl"
  height: Scalars["Int"]["output"]
  src: Scalars["String"]["output"]
  srcSet: Scalars["String"]["output"]
  url: Scalars["String"]["output"]
  width: Scalars["Int"]["output"]
}
// ...
export type Image = {
  __typename?: "Image"
  aspectRatio: Scalars["Float"]["output"]
  blurhash?: Maybe<Scalars["String"]["output"]>
  caption?: Maybe<Scalars["String"]["output"]>
  cropped?: Maybe<CroppedImageUrl>
  deepZoom?: Maybe<DeepZoom>
  geminiToken?: Maybe<Scalars["String"]["output"]>
  height?: Maybe<Scalars["Int"]["output"]>
  href?: Maybe<Scalars["String"]["output"]>
  imageURL?: Maybe<Scalars["String"]["output"]>
  imageVersions?: Maybe<Array<Maybe<Scalars["String"]["output"]>>>
  internalID?: Maybe<Scalars["ID"]["output"]>
  isDefault?: Maybe<Scalars["Boolean"]["output"]>
  isZoomable?: Maybe<Scalars["Boolean"]["output"]>
  maxTiledHeight?: Maybe<Scalars["Int"]["output"]>
  maxTiledWidth?: Maybe<Scalars["Int"]["output"]>
  orientation?: Maybe<Scalars["String"]["output"]>
  originalHeight?: Maybe<Scalars["Int"]["output"]>
  originalWidth?: Maybe<Scalars["Int"]["output"]>
  placeholder?: Maybe<Scalars["String"]["output"]>
  position?: Maybe<Scalars["Int"]["output"]>
  resized?: Maybe<ResizedImageUrl>
  tileBaseURL?: Maybe<Scalars["String"]["output"]>
  tileFormat?: Maybe<Scalars["String"]["output"]>
  tileSize?: Maybe<Scalars["Int"]["output"]>
  title?: Maybe<Scalars["String"]["output"]>
  url?: Maybe<Scalars["String"]["output"]>
  versions?: Maybe<Array<Maybe<Scalars["String"]["output"]>>>
  width?: Maybe<Scalars["Int"]["output"]>
}
// ...
export type Node = {
  id: Scalars["ID"]["output"]
}
// ...
export type Partner = Node & {
  __typename?: "Partner"
  alertsConnection?: Maybe<PartnerAlertsConnection>
  allArtistsConnection?: Maybe<ArtistPartnerConnection>
  analytics?: Maybe<AnalyticsPartnerStats>
  articlesConnection?: Maybe<ArticleConnection>
  artistsConnection?: Maybe<ArtistPartnerConnection>
  artistsSearchConnection?: Maybe<ArtistConnection>
  artistsWithAlertCountsConnection?: Maybe<ArtistsWithAlertCountsConnection>
  artworkImportsConnection?: Maybe<ArtworkImportConnection>
  artworksConnection?: Maybe<ArtworkConnection>
  artworksSearchConnection?: Maybe<ArtworkConnection>
  cached?: Maybe<Scalars["Int"]["output"]>
  categories?: Maybe<Array<Maybe<PartnerCategory>>>
  cities?: Maybe<Array<Maybe<Scalars["String"]["output"]>>>
  claimed?: Maybe<Scalars["Boolean"]["output"]>
  collectingInstitution?: Maybe<Scalars["String"]["output"]>
  contact?: Maybe<Contact>
  contactsConnection?: Maybe<ContactConnection>
  counts?: Maybe<PartnerCounts>
  defaultProfileID?: Maybe<Scalars["String"]["output"]>
  displayArtistsSection?: Maybe<Scalars["Boolean"]["output"]>
  displayFullPartnerPage?: Maybe<Scalars["Boolean"]["output"]>
  displayWorksSection?: Maybe<Scalars["Boolean"]["output"]>
  distinguishRepresentedArtists?: Maybe<Scalars["Boolean"]["output"]>
  documentsConnection?: Maybe<PartnerDocumentConnection>
  featuredKeywords: Array<Scalars["String"]["output"]>
  featuredShow?: Maybe<Show>
  filterArtworksConnection?: Maybe<FilterArtworksConnection>
  /** @deprecated Prefer displayFullPartnerPage */
  fullProfileEligible?: Maybe<Scalars["Boolean"]["output"]>
  hasFairPartnership?: Maybe<Scalars["Boolean"]["output"]>
  hasVisibleFollowsCount: Scalars["Boolean"]["output"]
  href?: Maybe<Scalars["String"]["output"]>
  id: Scalars["ID"]["output"]
  initials?: Maybe<Scalars["String"]["output"]>
  inquiryRequest?: Maybe<PartnerInquiryRequest>
  internalID: Scalars["ID"]["output"]
  isDefaultProfilePublic?: Maybe<Scalars["Boolean"]["output"]>
  isInquireable: Scalars["Boolean"]["output"]
  isLinkable?: Maybe<Scalars["Boolean"]["output"]>
  isPreQualify?: Maybe<Scalars["Boolean"]["output"]>
  isVerifiedSeller?: Maybe<Scalars["Boolean"]["output"]>
  location?: Maybe<Location>
  /** @deprecated Prefer to use `locationsConnection`. [Will be removed in v2] */
  locations?: Maybe<Array<Maybe<Location>>>
  locationsConnection?: Maybe<LocationConnection>
  merchantAccount?: Maybe<PartnerMerchantAccount>
  meta?: Maybe<PartnerMeta>
  name?: Maybe<Scalars["String"]["output"]>
  partnerAlertHitsConnection?: Maybe<PartnerAlertHitsConnection>
  partnerPageEligible?: Maybe<Scalars["Boolean"]["output"]>
  partnerType?: Maybe<Scalars["String"]["output"]>
  profile?: Maybe<Profile>
  profileArtistsLayout?: Maybe<Scalars["String"]["output"]>
  profileBannerDisplay?: Maybe<Scalars["String"]["output"]>
  showPromoted?: Maybe<Scalars["Boolean"]["output"]>
  showsConnection?: Maybe<ShowConnection>
  showsSearchConnection?: Maybe<ShowConnection>
  slug: Scalars["ID"]["output"]
  type?: Maybe<Scalars["String"]["output"]>
  vatInformation?: Maybe<Scalars["String"]["output"]>
  vatNumber?: Maybe<Scalars["String"]["output"]>
  vatStatus?: Maybe<Scalars["String"]["output"]>
  viewingRoomsConnection?: Maybe<ViewingRoomsConnection>
  website?: Maybe<Scalars["String"]["output"]>
}
// ...
export type PartnerAlertsConnection = {
  __typename?: "PartnerAlertsConnection"
  edges?: Maybe<Array<Maybe<PartnerAlertsEdge>>>
  pageCursors: PageCursors
  pageInfo: PageInfo
  totalCount?: Maybe<Scalars["Int"]["output"]>
}
// ...
export type TopArtworksToolQuery = {
  __typename?: "Query"
  partner?: {
    __typename?: "Partner"
    name?: string | null
    analytics?: {
      __typename?: "AnalyticsPartnerStats"
      rankedStats?: {
        __typename?: "AnalyticsRankedStatsConnection"
        edges?: Array<{
          __typename?: "AnalyticsRankedStatsEdge"
          node?: {
            __typename?: "AnalyticsRankedStats"
            value: number
            entity?:
              | { __typename: "Artist" }
              | {
                  __typename: "Artwork"
                  id: string
                  slug: string
                  title?: string | null
                  isUnlisted: boolean
                  artist?: {
                    __typename?: "Artist"
                    name?: string | null
                  } | null
                  image?: {
                    __typename?: "Image"
                    cropped?: {
                      __typename?: "CroppedImageUrl"
                      url: string
                    } | null
                  } | null
                }
              | { __typename: "Show" }
              | { __typename: "ViewingRoom" }
              | null
          } | null
        } | null> | null
      } | null
    } | null
  } | null
}